authority validation functionality moved to httpx package
